use-me: refactor to handle invalidation and login/logout better
Now, use-me takes into account the state of the auth context, while
still caching data.

I think we should likely merge this hook into the auth context
/somehow/, but I am yet to put enough thought into it.  For now, we have
an API that consists of an Axios instance that is "magically"
authenticated and a hook that allows us to access information about the
current user.   This is a fairly flexible configuration, so we can think
of better solutions later.